Moving BOSH to data-access-layer folder
