+++ api-controllers/FabrikBaseController.js
@@ -8,7 +8,7 @@ const config = require('../common/config');
 const utils = require('../broker/lib/utils');
 const errors = require('../common/errors');
 const cf = require('../broker/lib/cf');
-const bosh = require('../broker/lib/bosh');
+const bosh = require('../data-access-layer/bosh');
 const fabrik = require('../broker/lib/fabrik');
 const backupStore = require('../broker/lib/iaas').backupStore;
 const catalog = require('../broker/lib/models/catalog');
