+++ api-controllers/ServiceFabrikAdminController.js
@@ -7,12 +7,12 @@ const catalog = require('../broker/lib/models/catalog');
 const errors = require('../common/errors');
 const logger = require('../common/logger');
 const config = require('../common/config');
-const NetworkSegmentIndex = require('../broker/lib/bosh/NetworkSegmentIndex');
+const NetworkSegmentIndex = require('../data-access-layer/bosh/NetworkSegmentIndex');
 const backupStore = require('../broker/lib/iaas').backupStore;
 const FabrikBaseController = require('./FabrikBaseController');
 const utils = require('../broker/lib/utils');
 const fabrik = require('../broker/lib/fabrik');
-const bosh = require('../broker/lib/bosh');
+const bosh = require('../data-access-layer/bosh');
 const ScheduleManager = require('../broker/lib/jobs');
 const BackupReportManager = require('../broker/lib/reports');
 const CONST = require('../common/constants');
