+++ broker/lib/fabrik/DBManager.js
@@ -6,7 +6,7 @@ const config = require('../config');
 const logger = require('../logger');
 const catalog = require('../models/catalog');
 const DirectorManager = require('./DirectorManager');
-const bosh = require('../bosh');
+const bosh = require('../../../data-access-layer/bosh');
 const utils = require('../utils');
 const errors = require('../errors');
 const ServiceBindingNotFound = errors.ServiceBindingNotFound;
