+++ broker/lib/fabrik/DirectorManager.js
@@ -6,7 +6,7 @@ const Promise = require('bluebird');
 const config = require('../config');
 const logger = require('../logger');
 const errors = require('../errors');
-const bosh = require('../bosh');
+const bosh = require('../../../data-access-layer/bosh');
 const cf = require('../cf');
 const backupStore = require('../iaas').backupStore;
 const utils = require('../utils');
