+++ broker/lib/fabrik/FabrikStatusPoller.js
@@ -7,7 +7,7 @@ const catalog = require('../models/catalog');
 const pubsub = require('pubsub-js');
 const logger = require('../logger');
 const ServiceFabrikOperation = require('./ServiceFabrikOperation');
-const bosh = require('../bosh');
+const bosh = require('../../../data-access-layer/bosh');
 const utils = require('../utils');
 const errors = require('../errors');
 const config = require('../config');
