+++ broker/lib/fabrik/OobBackupManager.js
@@ -7,7 +7,7 @@ const backupStore = require('../iaas').backupStoreForOob;
 const Agent = require('./Agent');
 const utils = require('../utils');
 const errors = require('../errors');
-const bosh = require('../bosh');
+const bosh = require('../../../data-access-layer/bosh');
 const NotFound = errors.NotFound;
 const UnprocessableEntity = errors.UnprocessableEntity;
 const CONST = require('../constants');
