+++ broker/lib/fabrik/VirtualHostInstance.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const BaseInstance = require('./BaseInstance');
-const bosh = require('../bosh');
+const bosh = require('../../../data-access-layer/bosh');
 const utils = require('../utils');
 const mapper = require('./VirtualHostRelationMapper');
 const catalog = require('../models').catalog;
