+++ broker/lib/fabrik/VirtualHostRelationMapper.js
@@ -2,7 +2,7 @@
 
 const Promise = require('bluebird');
 const logger = require('../logger');
-const bosh = require('../bosh');
+const bosh = require('../../../data-access-layer/bosh');
 const cloudController = require('../cf').cloudController;
 const virtualHostStore = require('../iaas').virtualHostStore;
