+++ broker/lib/index.js
@@ -8,7 +8,7 @@ exports.logger = require('./logger');
 exports.utils = require('./utils');
 exports.store = require('./store');
 exports.middleware = require('./middleware');
-exports.bosh = require('./bosh');
+//exports.bosh = require('./bosh');
 exports.cf = require('./cf');
 exports.models = require('./models');
 //exports.routes = require('./routes');
