+++ broker/lib/jobs/BnRStatusPollerJob.js
@@ -10,7 +10,7 @@ const logger = require('../logger');
 const errors = require('../errors');
 const DeploymentAlreadyLocked = errors.DeploymentAlreadyLocked;
 const config = require('../config');
-const bosh = require('../bosh');
+const bosh = require('../../../data-access-layer/bosh');
 const catalog = require('../models').catalog;
 const BackupService = require('../../../managers/backup-manager');
 const eventmesh = require('../../../eventmesh');
