+++ broker/lib/jobs/ScheduledOobDeploymentBackupJob.js
@@ -10,7 +10,7 @@ const CONST = require('../constants');
 const errors = require('../errors');
 const NotFound = errors.NotFound;
 const backupStore = require('../iaas').backupStoreForOob;
-const bosh = require('../bosh');
+const bosh = require('../../../data-access-layer/bosh');
 const ScheduleManager = require('./ScheduleManager');
 
 class ScheduledOobDeploymentBackupJob extends BaseJob {
