+++ data-access-layer/bosh/BoshDirectorClient.js
@@ -4,14 +4,14 @@ const parseUrl = require('url').parse;
 const Promise = require('bluebird');
 const _ = require('lodash');
 const yaml = require('js-yaml');
-const errors = require('../errors');
+const errors = require('../../common/errors');
 const Timeout = errors.Timeout;
-const logger = require('../logger');
-const utils = require('../utils');
-const CONST = require('../constants');
+const logger = require('../../common/logger');
+const utils = require('../../broker/lib/utils');
+const CONST = require('../../common/constants');
 const retry = utils.retry;
 const HttpClient = utils.HttpClient;
-const config = require('../config');
+const config = require('../../common/config');
 const NotFound = errors.NotFound;
 const BadRequest = errors.BadRequest;
