+++ data-access-layer/bosh/manifest/Network.js
