+++ data-access-layer/bosh/manifest/Networks.js
