+++ managers/backup-manager/BackupService.js
@@ -5,7 +5,7 @@ const Promise = require('bluebird');
 const config = require('../../broker/lib/config');
 const logger = require('../../common/logger');
 const errors = require('../../common/errors');
-const bosh = require('../../broker/lib/bosh');
+const bosh = require('../../data-access-layer/bosh');
 const backupStore = require('../../broker/lib/iaas').backupStore;
 const utils = require('../../broker/lib/utils');
 const eventmesh = require('../../eventmesh');
