+++ test/test_broker/bosh.BoshDirectorClient.spec.js
@@ -9,7 +9,7 @@ const config = require('../../broker/lib/config');
 const logger = require('../../broker/lib/logger');
 const NotFound = errors.NotFound;
 const BadRequest = errors.BadRequest;
-const BoshDirectorClient = require('../../broker/lib/bosh/BoshDirectorClient');
+const BoshDirectorClient = require('../../data-access-layer/bosh/BoshDirectorClient');
 const yaml = require('js-yaml');
 const id = uuid.v4();
 const deployment_name = id;
