+++ test/test_broker/bosh.EvaluationContext.spec.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const proxyquire = require('proxyquire');
-const EvaluationContext = proxyquire('../../broker/lib/bosh/EvaluationContext', {
+const EvaluationContext = proxyquire('../../data-access-layer/bosh/EvaluationContext', {
   crypto: {
     randomBytes: function (n) {
       return _.fill(new Buffer(n), 120);
