+++ test/test_broker/fabrik.DBManager.spec.js
@@ -3,7 +3,7 @@
 const proxyquire = require('proxyquire');
 const Promise = require('bluebird');
 const _ = require('lodash');
-const BoshDirectorClient = require('../../broker/lib/bosh/BoshDirectorClient');
+const BoshDirectorClient = require('../../data-access-layer/bosh/BoshDirectorClient');
 const errors = require('../../broker/lib/errors');
 const config = require('../../broker/lib/config');
 const utils = require('../../broker/lib/utils');
