+++ test/test_broker/fabrik.FabrikStatusPoller.spec.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const proxyquire = require('proxyquire');
 const logger = require('../../broker/lib/logger');
 const DirectorManager = lib.fabrik.DirectorManager;
-const BoshDirectorClient = lib.bosh.BoshDirectorClient;
+const BoshDirectorClient = require('../../data-access-layer/bosh').BoshDirectorClient;
 const CONST = require('../../broker/lib/constants');
 const errors = require('../../broker/lib/errors');
 const ServiceFabrikClient = require('../../broker/lib/cf/ServiceFabrikClient');
