+++ test/test_broker/mocks/director.js
@@ -8,7 +8,7 @@ const lib = require('../../../broker/lib');
 const CONST = require('../../../broker/lib/constants');
 const agent = require('./agent');
 const config = lib.config;
-const bosh = lib.bosh;
+const bosh = require('../../../data-access-layer/bosh');
 const NetworkSegmentIndex = bosh.NetworkSegmentIndex;
 const prefix = 'service-fabrik';
 const networkSegmentIndex = 21;
